Move utils to correct module
