+++ cg/services/order_validation_service/workflows/tomte/validation/field/tomte_model_validator.py
@@ -1,8 +1,10 @@
 from pydantic_core import ValidationError
 
 from cg.services.order_validation_service.models.errors import ValidationErrors
-from cg.services.order_validation_service.utils import convert_errors
 from cg.services.order_validation_service.workflows.tomte.models.order import TomteOrder
+from cg.services.order_validation_service.workflows.tomte.validation.field.utils import (
+    convert_errors,
+)
 
 
 class TomteModelValidator:
